finished userCreatesBuildStep test and start userRunBuildWithStep test
