+++ src/main/java/com/example/teamcity/ui/elements/BuildRunElements.java
@@ -1,0 +1,18 @@
+package com.example.teamcity.ui.elements;
+
+import com.codeborne.selenide.SelenideElement;
+import lombok.Getter;
+
+@Getter
+public class BuildRunElements extends  BasePageElement{
+    private  SelenideElement buildStatus;
+
+
+    public BuildRunElements(SelenideElement element) {
+        //this.buildStatus=
+        super(element);
+    }
+
+
+
+}
